LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focusgroup/tentative/backward-navigation/descends-into-extending-focusgroup-expected.txt

item1
item2 item3
item4

FAIL When the focus is set on the last item of a focusgroup and the previous item is a descendant of a subtree, a backward arrow key press should move the focus to that previous item within the subtree. assert_equals: expected Element node <span id="item3" tabindex="-1">item3</span> but got Element node <span id="item4" tabindex="-1">item4</span>
